+++ src/gitstore/__init__.py
@@ -1,15 +1,18 @@
-from .repo import GitStore, ReflogEntry
+from .repo import GitStore, RefDict, ReflogEntry, Signature
 from .mirror import MirrorDiff, RefChange
 from .notes import NoteDict, NoteNamespace, NotesBatch
 from .fs import FS, WriteEntry, retry_write
-from .tree import WalkEntry
+from .tree import BlobOid, GitError, WalkEntry
+from .batch import Batch
 from .exceptions import StaleSnapshotError
 from .copy import ChangeReport, ChangeAction, ChangeError, FileEntry, FileType, disk_glob
 
 __all__ = [
-    "GitStore", "ReflogEntry", "MirrorDiff", "RefChange",
+    "GitStore", "RefDict", "ReflogEntry", "Signature",
+    "MirrorDiff", "RefChange",
     "NoteDict", "NoteNamespace", "NotesBatch",
     "FS", "WriteEntry", "retry_write", "StaleSnapshotError",
+    "Batch", "BlobOid", "GitError",
     "ChangeReport", "ChangeAction", "ChangeError", "FileEntry", "FileType", "disk_glob",
     "WalkEntry",
 ]